+++ tests/TestCase.php
@@ -2,6 +2,7 @@
 
 namespace Spatie\Fractal\Test;
 
+use Illuminate\Support\Facades\Route;
 use Spatie\Fractal\FractalServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
@@ -44,6 +45,8 @@ abstract class TestCase extends Orchestra
                 'publisher' => 'Bloody Fantasy inc.',
             ],
         ];
+
+        $this->setupRoutes();
     }
 
     protected function getPackageProviders($app)
@@ -63,5 +66,15 @@ abstract class TestCase extends Orchestra
         if ($this->defaultSerializer != '') {
             $app['config']->set('fractal.default_serializer', $this->defaultSerializer);
         }
+    }
+
+    protected function setupRoutes()
+    {
+        Route::get('auto-includes', function() {
+           return $this->fractal
+               ->collection($this->testBooks)
+               ->transformWith(TestTransformer::class)
+               ->toArray();
+        });
     }
 }